Workaround for T-M register interdependencies
T and M are clocked with different clocks derived from the same strob.
Problem is that this is enough for data to come too early or too late
and cause various misbehaviors.
To solve that strob1 (STROB*FP) has been separated from the register
enable signals and T/M registers are now driven that way.